Fixed bug with dtype
Channels that were held constant while their parent clockline ramped another channel did not preserve the dtype of the value. Instead of forcing the value to float32 always, we now respect the dtype stored for the channel.

+++ labscript.py
@@ -1051,7 +1051,7 @@ class Output(Device):
                         if ((outarray<self.limits[0])|(outarray>self.limits[1])).any():
                             raise LabscriptError('The function %s called on "%s" at t=%d generated a value which falls outside the base unit limits (%d to %d)'%(self.timeseries[i]['function'],self.name,midpoints[0],limits[0],limits[1]))
                 else:
-                    outarray = empty(len(time),dtype=float32)
+                    outarray = empty(len(time),dtype=self.dtype)
                     outarray.fill(self.timeseries[i])
                 outputarray.append(outarray)
             else:
